requirements.txt
alembic==1.18.4
APScheduler==3.11.2
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
email-validator==2.3.0
et_xmlfile==2.0.0
Flask==2.3.3
Flask-APScheduler==1.12.4
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-Reuploaded==1.0.0
Flask-SQLAlchemy==3.0.5
Flask-Uploads==0.2.1
Flask-WTF==1.1.1
greenlet==3.3.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
openpyxl==3.1.2
pillow==12.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
reportlab==4.0.4
requests==2.31.0
six==1.17.0
SQLAlchemy==2.0.48
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
Werkzeug==2.3.7
WTForms==3.0.1
